+++ apps/consent-ui/src/components/Layouts/PageLayout/PageLayout.module.scss
